Add google analytics.
This adds google analytics, and parameterizes the UA
so that we don't pollute production with metrics
logged in other environments.

The default UA is that for staging / development / any
non live environment.

I also added a flag, `--dev`, so that the server can
be ran locally in a mode where it reloads the template
with each request (to allow for more rapid local
development).