Limit clients to 1 request per second.
Every now and then someone writes a bot to scrape the API that causes a
bunch of 5XXs. This limit is a bit aggressive, particularly given it
applies to both requests against the API and UI, but it'll at the very
least slow bots down to the point where they shouldn't be able to cause
a disruption.

If we want to make this more robust in the future we should add a
reverse proxy and rate limit the API and UI endpoints separately.